add hc support for wildboottest()
